move Header and Footer to layouts folder
