+++ layouts/Layout.tsx
@@ -1,8 +1,8 @@
 import React, { useState, ReactNode } from 'react';
 import styled from 'styled-components';
 
-import Header from '../components/Header/Header';
-import Footer from '../components/Footer/Footer';
+import Header from './Header/Header';
+import Footer from './Footer/Footer';
 import Button from '../components/core/Button/Button';
 import useScrollPosY from '../components/hooks/useScrollPosY';
